update block sync

+++ back-end-projects/OntSynHandler/sql/tbl_oep5.sql
@@ -18,13 +18,13 @@ SET FOREIGN_KEY_CHECKS = 0;
 DROP TABLE IF EXISTS `tbl_oep5`;
 CREATE TABLE `tbl_oep5`
 (
-    `contract_hash` varchar(255)   NOT NULL DEFAULT '' COMMENT '合约hash值',
-    `name`          varchar(255)   NOT NULL DEFAULT '' COMMENT 'OEP5代币名称',
-    `total_supply`  decimal(15, 0) NOT NULL COMMENT 'OEP5代币总量',
-    `symbol`        varchar(255)   NOT NULL DEFAULT '' COMMENT 'OEP5代币符号',
-    `create_time`   datetime       NOT NULL COMMENT '创建时间，yyyy-MM-dd',
-    `audit_flag`    int(11)        NOT NULL COMMENT '审核标识，1：审核通过 0：未审核',
-    `update_time`   datetime                DEFAULT NULL COMMENT '更新时间，yyyy-MM-dd',
+    `contract_hash` varchar(255) NOT NULL DEFAULT '' COMMENT '合约hash值',
+    `name`          varchar(255) NOT NULL DEFAULT '' COMMENT 'OEP5代币名称',
+    `total_supply`  BIGINT       NOT NULL COMMENT 'OEP5代币总量',
+    `symbol`        varchar(255) NOT NULL DEFAULT '' COMMENT 'OEP5代币符号',
+    `create_time`   datetime     NOT NULL COMMENT '创建时间，yyyy-MM-dd',
+    `audit_flag`    int(11)      NOT NULL COMMENT '审核标识，1：审核通过 0：未审核',
+    `update_time`   datetime              DEFAULT NULL COMMENT '更新时间，yyyy-MM-dd',
     PRIMARY KEY (`contract_hash`)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8;
